Refactor and enhance Google Tag Manager - Move deployment scripts to 'build' folder. - Modify GoogleAnalyticsManager to use only a single config file. - Add post build script and configure npm to use it. - Make post build script create the Google Analytics config. - Move Google Analytics config to environment variables. - Break out cache buster to it's own CacheUtil file. - Refactor to use CacheUtil file. - Fix bug in CacheUtil where URL fragment wasn't encoded correctly.
