+++ package.json
@@ -5,7 +5,7 @@
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
+    "build": "webpack && node ./build/post-build.js"
   },
   "repository": {
     "type": "git",
