--- wwwroot/config/googleAnalytics.example.json
@@ -1,3 +1,0 @@
-{
-    "key": "KEY_HERE"
-}
