+++ wwwroot/config/googleAnalytics.json
@@ -1,3 +1,4 @@
 {
-    "useGoogleAnalytics": true
+    "useGoogleAnalytics": false,
+    "key": "G-WE86V4LDMM"
 }
