+++ wwwroot/modules/components/googleAnalyticsManager.js
@@ -1,3 +1,5 @@
+import CacheUtil from '../util/cacheUtil.js';
+
 export default class GoogleAnalyticsManager {
 
 
@@ -28,23 +30,12 @@ export default class GoogleAnalyticsManager {
     async loadConfigAsync() {
         if (this.#config) return this.#config;
 
-        let global = {};
-        let instance = {};
-
-        let resp = await fetch('config/googleAnalytics.json');
+        const url = `./config/googleAnalytics.json${CacheUtil.getCacheBuster() ?? ''}`;
+        let resp = await fetch(url);
         if (resp.ok) {
-            global = JSON.parse(await resp.text());
-            if (global.useGoogleAnalytics) {
-                resp = await fetch('config/googleAnalytics.instance.json');
-                if (resp.ok) {
-                    instance = JSON.parse(await resp.text());
-                }
-            }
+            this.#config = JSON.parse(await resp.text());
         }
-        this.#config = {
-            ...global,
-            ...instance
-        }
+
         return this.#config;
     }
 
@@ -60,7 +51,7 @@ export default class GoogleAnalyticsManager {
 
             const script = document.createElement('script');
             script.async = true;
-            script.src = `https://www.googletagmanager.com/gtag/js?id=${config.key}`;
+            script.src = `https://www.googletagmanager.com/gtag/js?id=${encodeURIComponent(config.key)}`;
             script.onload = (ev) => {
                 window.dataLayer = window.dataLayer || [];
                 function gtag() { dataLayer.push(arguments); }
